Rework motion dashboard UX
This has a lot of changes, including some backend changes, to support
a better motion dashboard view.  The views now depend on state and also
render some better moderation tools for the relevant states.

Issue #7 Admin: Motion dashboard with state management
